# Reformat history table