--- a/crates/core/src/db/convert.rs
+++ b/crates/core/src/db/convert.rs
@@ -254,6 +254,7 @@ impl<'a, 'b> TryFrom<&'a Row<'b>> for ExchangeSummary {
     fn try_from(row: &'a Row<'b>) -> Result<Self, Self::Error> {
         Ok(Self {
             id: row.get("id")?,
+            recipe_id: row.get("recipe_id")?,
             profile_id: row.get("profile_id")?,
             start_time: row.get("start_time")?,
             end_time: row.get("end_time")?,