## Description

_Describe the change. If there is an associated issue, please include the issue link (e.g. "Closes #xxx"). For UI changes, please also include screenshots._

### Before

```
works	6ea637ea-0bc2-4f8e-8f18-13ece7a2e0fc	200	2025-02-26T18:39:07Z-05:00
works	0e71a5ed-69ac-49e3-b858-87fa88e9a4fa	200	2024-12-17T17:34:58Z-05:00
works	a4661a36-ab76-4ce0-8c8b-999b10033aa1	503	2024-12-11T15:15:56Z-05:00
works	59ba8ecf-6c50-4030-b327-319072763374	503	2024-12-11T15:15:55Z-05:00
works	87f7f7a2-8ba3-4dc2-b652-5582d8250302	200	2024-08-14T10:19:10Z-04:00
works	44c25873-64ae-42d2-88ec-cd96fba27c91	200	2024-08-14T10:19:07Z-04:00
works	4aa4c929-78d6-4bcb-bb74-13a5e9c6e502	200	2024-08-14T10:19:04Z-04:00
works	155e8f96-a503-40be-88b6-3e55e640a6d4	200	2024-08-04T15:13:49Z-04:00
works	245d5d1d-7bf6-4901-90be-d366b0a4c2ee	200	2024-07-06T10:03:40Z-04:00
```

### After

```
Recipe       Profile Time                       Status Request ID
upload_image works   2025-02-26T20:51:34Z-05:00 200    96fd2e98-94ef-4ef7-947a-a3f77b1b1627
get_image    works   2025-02-26T20:51:33Z-05:00 200    c18d69e0-3243-4f20-b629-41ebdc8461bb
modify_user  works   2025-02-26T20:51:32Z-05:00 200    f5bf3dfe-d0f8-4d86-ab31-2746e657d123
get_user     works   2025-02-26T20:51:32Z-05:00 200    735973d0-f521-4d52-b53e-6fc8ef4735f7
get_user     works   2025-02-26T20:51:31Z-05:00 200    6194df9b-cddc-417c-9ace-b792beee0eac
get_users    works   2025-02-26T20:51:30Z-05:00 200    77f578f7-f313-44df-af33-4ebd67a19a66
login        works   2025-02-26T18:38:36Z-05:00 200    419b8890-ca60-4bdd-825c-999909ec1180
```

## Known Risks

_What issues could potentially go wrong with this change? Is it a breaking change? What have you done to mitigate any potential risks?_

There's no handling of row overflow, so it could get pretty ugly

## QA

_How did you test this?_

Manually

## Checklist

- [x] Have you read `CONTRIBUTING.md` already?
- [x] Did you update `CHANGELOG.md`?
  - Only user-facing changes belong in the changelog. Internal changes such as refactors should only be included if they'll impact users, e.g. via performance improvement.
- [x] Did you remove all TODOs?
  - If there are unresolved issues, please open a follow-on issue and link to it in a comment so future work can be tracked